+++ src/pages/onboarding/ChoosePlan.js
@@ -6,7 +6,8 @@ import { showAlert } from 'src/actions/globalAlert';
 import { CenteredLogo, Loading, PlanPicker } from 'src/components';
 import Steps from './components/Steps';
 import { getPlans } from 'src/actions/account';
-import { getBillingCountries, billingCreate, updateSubscription } from 'src/actions/billing';
+import { getBillingCountries, updateSubscription } from 'src/actions/billing';
+import { billingCreate } from 'src/actions/billingCreate';
 import { selectVisiblePlans } from 'src/selectors/accountBillingInfo';
 import { changePlanInitialValues } from 'src/selectors/accountBillingForms';
 import PaymentForm from 'src/pages/billing/forms/fields/PaymentForm';
